+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed SWIMMING cost per level.
 - Fixed costs for characteristics with ADD_MODIFIERS_TO_BASE. [#412](https://github.com/dmdorman/hero6e-foundryvtt/issues/412)
 - Added RIDING discount for TRANSPORT_FAMILIARITY. [#397](https://github.com/dmdorman/hero6e-foundryvtt/issues/397)
+- Automations are immune to mental attacks. [#338](https://github.com/dmdorman/hero6e-foundryvtt/issues/338)
 
 
 # Version 3.0.44